Created CARLA enviorment for velocity data collection. Additionally added script to visuzalize the veloicty
Created carla script for velocity, in addition to visualizing the data collected at 10fps. Running script to collect numpy files that will be uploaded on the shared OneDrive. Car will stop for traffic cars and traffic lights, working with the walkers (pedestrians) has caused some issues, but I am able to spawn them in and have them move. Will be working on collecting more datasets with the walker logic intact.